Add analysis module
