+++ 1DGratingAnalysis.py
@@ -1,80 +1,10 @@
 import os
 import numpy as np
-import scipy.signal as sig
-import Functions.Maths as math
 import Functions.Organisation as org
+
 import src.fileIO as fileIO
+import src.analysis as anal
 
-
-def trim_to_region_of_interest(image,
-                               height,
-                               width):
-    '''
-    Trim numpy array to specified region of interest.
-    Args:
-        image: <np.array> array of pixels
-        height: <int> number of pixels in vertical axis
-        width: <int> number of pixels in horizontal axis
-    Returns:
-        pixels: <array> pixel array of image
-        regionofinterest: <array> trimmed pixel array to region of interest
-    '''
-    region_of_interest = image[0: height, 0: width]
-    return region_of_interest
-
-
-def calculate_distanceperpixel(distance_value, distance_unit, number_of_pixels):
-    '''
-    Calaculate the distance each pixel represents in SEM image. Distance is
-    set to micrometer scale, i.e. if image file returns um as unit, the scalar
-    returns 1, mm is a 1000x more and nm is 1000x less. Pulls data from params
-    dictionary and calculates from array. Ensure arguments are individual array
-    elements.
-    Args:
-        distance: <string> of the form '123nm'
-        number_of_pixels: <int>
-    Returns:
-        distanceperpixel: <float> distance in um per pixel
-    '''
-    unit_scalar_map = {
-        'mm': 1E3,
-        'um': 1,
-        'nm': 1E-3}
-    scalar = unit_scalar_map[distance_unit]
-    distanceperpixel = (distance_value * scalar) / int(number_of_pixels)
-    return distanceperpixel
-
-
-def FourierTransformRow(row,
-                        num_peaks,
-                        distance_pixel):
-    '''
-    Calculate fourier transform of image row.
-    Args:
-        row: <array> pixel values for image row
-        num_peaks: <int> number of peaks to pull from fourier transform (number
-                    of periods to analyse)
-        distance_pixel: <float> distance in um per pixel
-    Returns:
-        frequencies: <array> fourier space frequency values of period peaks
-        periods: <array> signal periods from fourier transform converted from um
-                    to nm
-    '''
-    samplesize = len(row)
-    xspace = range(0, samplesize, 1)
-    fspace = np.fft.rfftfreq(len(xspace), 1)
-    fft = np.fft.rfft(row)
-    magnitude = np.abs(fft)
-    peaks, _ = sig.find_peaks(x=magnitude)
-    prominences, _, _ = sig.peak_prominences(
-        x=magnitude,
-        peaks=peaks)
-    indices = np.array(np.argsort(prominences)[-num_peaks:][::-1])
-    periodpeaks = peaks[indices]
-    fsteps = [p / (distance_pixel * samplesize) for p in periodpeaks]
-    periods = [(1 / f) * 1E3 for f in fsteps]
-    frequencies = fspace[indices]
-    return frequencies, periods
 
 
 def dictionary_union(primary_dict,
@@ -122,11 +52,11 @@ if __name__ == '__main__':
 
             image = fileIO.read_image(file_path=imagepath)
 
-            grating_region = trim_to_region_of_interest(
+            grating_region = anal.trim_to_region_of_interest(
                 image=image,
                 height=sem_parameters['image_height'],
                 width=sem_parameters['image_width'])
-            distanceperpixel = calculate_distanceperpixel(
+            distanceperpixel = anal.calculate_distanceperpixel(
                 distance_value=sem_parameters['calibration_distance_value'],
                 distance_unit=sem_parameters['calibration_distance_unit'],
                 number_of_pixels=sem_parameters['calibration_number_of_pixels'])
@@ -134,7 +64,7 @@ if __name__ == '__main__':
             periods = []
             frequencies = []
             for i in range(len(grating_region)):
-                fourierfrequencies, fourierperiods = FourierTransformRow(
+                fourierfrequencies, fourierperiods = anal.FourierTransformRow(
                     row=grating_region[i],
                     num_peaks=3,
                     distance_pixel=distanceperpixel)
@@ -148,13 +78,13 @@ if __name__ == '__main__':
                         np.sum(p) / len(p)
                         for p in np.array(periods).T],
                     'Period_Errors_nm': [
-                        math.StandardErrorMean(x=p)
+                        anal.StandardErrorMean(x=p)
                         for p in np.array(periods).T],
                     'Average_Frequencies': [
                         np.sum(f) / len(f)
                         for f in np.array(frequencies).T],
                     'Frequencies_Errors': [
-                        math.StandardErrorMean(x=f)
+                        anal.StandardErrorMean(x=f)
                         for f in np.array(frequencies).T]})
 
             fileIO.save_json(
